Fix more null pointers
Fixed more null pointer exceptions with logging
Added method to get the robot pose. Currently just
using odometry, can be changed to use pose
estimation if that is something we decide to
do
Added a method for zeroing the odometry. Helpful for
testing that the odometry actually works
Removed not logging on disable, because having
information is always useful