Add PathFollowState
Added a state to follow a given trajectory
Has a few more options than last year, and many more
constructors
You can now specify if you want the path to be time
stopped, position stopped,
or both. If both are true, it will ensure both end
conditions are true, otherwise, it will just check for the one that was selected
Has many constructors for as generic, or as detailed as you
need it to be.
Has methods for setting both initialization (odometry setup)
and end stopping.
For end stopping, a position delta is also required, so the
program knows what is good enough.
Added constants for some of the defaults.
PIDControllers still to be tuned.