Add precomputation for bounds
Added a precomputation to figure out dynamic boinds for path finding
based on the navgrid given by pathplanner.
Its a bit slow, and there are overlaps, but it works well, and does its
job of reducing the number of bounding boxes to avoid.
It also outputs everything to a text file that is super easily
copy/pasted into code.