Update following with events
THIS IS A BREAKING CHANGE
STUFF MIGHT NOT WORK.
DEBUGGING COMMITS WILL HAPPEN, AND PROBABLY SOON
Changed the way path following with events works to now allow for event
markers to be placed at points at which you wish to pause, and do
something else.
Updated the path class to allow for this. I did need to parse the JSON
myself to get this to work, but it is fine.
Updated the path following to automatically add the pause points put in
the path
Still keeping the ability to add new pause points in code though
Updated the path following to automatically unpause when the pause
command finishes running