Set sensor output range
This is the only way I can think of to set the output range.
There is no method accessible from the absolute encoder, there is no
method on the PID controller, nothing.
So, I decided I was going to set it myself.
Using reflection, I can pull the data I need, and using the JNI I can
set what I need to.
Still doesn't work. We will figure it out, but I am still committing
this code because I want to work on it more.