Add teleop gyro reset
Added resetting of the gyro at the start of teleop.
If on red, the gyro is 180 degrees off, so rotate it back
This is because of the fact that we need a unified coordinate system,
and that causes the angle of the gyro to be 180 degrees off when on red
side. This just resets the gyro to be better for driveteam